+++ 3rd_Deliverable/app_code/app/controllers/__init__.py
@@ -1,0 +1,7 @@
+from app.controllers.auth_controller import AuthController
+from app.controllers.club_mainpage_controller import ClubMainPageController
+from app.controllers.home_page_controller import HomePageController
+from app.controllers.login_controller import LoginController
+from app.controllers.main_window import MainWindow
+from app.controllers.res_details_controller import ReservationDetailsPage
+from app.controllers.view_res_controller import ViewReservationsController
