+++ 3rd_Deliverable/app_code/app/controllers/home_page_controller.py
@@ -1,7 +1,7 @@
 from PySide6.QtCore import QObject, Signal
-from app.views.customer_homepage import CustomerHomePage
-from app.controllers.view_res_controller import ViewReservationsController
-from app.controllers.club_mainpage_controller import ClubMainPageController
+from app.views import CustomerHomePage
+from app.controllers import ViewReservationsController
+from app.controllers import ClubMainPageController
 from app.models import User, Role, Club
 from app.utils.container import Container
