+++ 3rd_Deliverable/app_code/app/views/__init__.py
@@ -1,0 +1,5 @@
+from app.views.customer_club_mainpage import CustomerClubMainPage
+from app.views.customer_homepage import CustomerHomePage
+from app.views.customer_view_res import CustomerViewReservations
+from app.views.login_view import LoginView
+from app.views.reservation_details_page import ReservationDetailsPage
