+++ 3rd_Deliverable/app_code/app/views/customer_homepage.py
@@ -4,7 +4,7 @@ from PySide6.QtWidgets import (
 )
 from PySide6.QtGui import QFont
 from PySide6.QtCore import Qt, QSize, Signal
-from app.models.club import Club
+from app.models import Club
 
 
 class CustomerHomePage(QWidget):
